Move plugin compliation to setup.py
